Move isGoodDefaultDomainQuery to its own package to prevent cyclic deps
