+++ client/package.json
@@ -30,6 +30,7 @@
 		"@automattic/data-stores": "workspace:^",
 		"@automattic/design-picker": "workspace:^",
 		"@automattic/domain-picker": "workspace:^",
+		"@automattic/domain-utils": "workspace:^",
 		"@automattic/explat-client": "workspace:^",
 		"@automattic/explat-client-react-helpers": "workspace:^",
 		"@automattic/format-currency": "workspace:^",
