+++ packages/data-stores/src/onboard/selectors.ts
@@ -1,6 +1,5 @@
-import { isGoodDefaultDomainQuery } from '@automattic/domain-picker';
+import { isGoodDefaultDomainQuery } from '@automattic/domain-utils';
 import type { State } from './reducer';
-
 export const getIsRedirecting = ( state: State ) => state.isRedirecting;
 export const getPlanProductId = ( state: State ) => state.planProductId;
 export const getLastLocation = ( state: State ) => state.lastLocation;
