+++ packages/domain-picker/src/index.tsx
@@ -9,10 +9,6 @@ export {
 
 export type { SUGGESTION_ITEM_TYPE } from './components/domain-suggestion-item';
 
-export {
-	mockDomainSuggestion,
-	isGoodDefaultDomainQuery,
-	getDomainSuggestionsVendor,
-} from './utils';
+export { mockDomainSuggestion, getDomainSuggestionsVendor } from './utils';
 export { useDomainSuggestions } from './hooks';
 export { DOMAIN_QUERY_MINIMUM_LENGTH } from './constants';
