+++ packages/domain-utils/jest.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	preset: '../../test/packages/jest-preset.js',
+};
